add users list to chat settings if it's a group chat; remove editing chat name if it's single user chat;
