+++ src/actions/chat.ts
@@ -92,9 +92,13 @@ export const getMessages = chatId => async (dispatch, getState) => {
             url: `${MESSAGE_URL + chatId}?timestamp=${timestamp || 0}`,
             method: 'get',
         })
+        const chatUsers = await doJsonAuthRequest({
+            url: `${CHAT_URL}/${chatId}`,
+            method: 'get',
+        })
         dispatch({
             type: LOAD_MESSAGES,
-            payload: { newMessages, chatId },
+            payload: { newMessages, chatId, chatUsers },
         })
         dispatch({ type: MESSAGES_REFRESH_END })
     } catch (e) {
@@ -199,7 +203,6 @@ export const updateChatSettings = chat => async dispatch => {
             method: 'put',
             data: { chatId: chat.chatId, chatName: newChatName, chatImage: newChatImage },
         })
-        dispatch(setActiveChat(newChat))
         dispatch({
             type: UPDATE_CHAT,
             payload: newChat,
