+++ src/components/Chat/index.tsx
@@ -9,20 +9,8 @@ import Icon from 'react-native-vector-icons/FontAwesome5'
 import { MessagesList } from './MessagesList'
 import { goToAuth } from '../../navigation/navigation'
 import { WHITE_COLOR, SOFT_BLUE_COLOR } from '../../helpers/styleConstants'
-import {
-    setFindMessagesInputValue,
-    cleanFindMessagesInputValue,
-    togglePlayer,
-    downloadPlayer,
-} from '../../actions/messages'
-import {
-    sendMessage,
-    unsetActiveChat,
-    getChatlistTimestamp,
-    openSearchBar,
-    closeSearchBar,
-    getMessages,
-} from '../../actions/chat'
+import { setFindMessagesInputValue, cleanFindMessagesInputValue, togglePlayer, downloadPlayer } from '../../actions/messages'
+import { sendMessage, unsetActiveChat, getChatlistTimestamp, openSearchBar, closeSearchBar, getMessages } from '../../actions/chat'
 import _ from 'lodash'
 
 const { height } = Dimensions.get('window') // it's for IphoneX
@@ -154,29 +142,15 @@ class Chat extends React.PureComponent<IProps, IState> {
                 />
                 {chat.isSearchBarActive ? (
                     <SearchMessagesBar>
-                        <Icon.Button
-                            size={20}
-                            onPress={() => this.nextMessage(currentMessageNumber)}
-                            backgroundColor="d6efef"
-                            name="arrow-up"
-                            color={SOFT_BLUE_COLOR}
-                        />
+                        <Icon.Button size={20} onPress={() => this.nextMessage(currentMessageNumber)} backgroundColor="d6efef" name="arrow-up" color={SOFT_BLUE_COLOR} />
                         <Text style={{ marginRight: 10, marginLeft: 10 }}>
                             {/* to do: refactor this code */}
                             {filteredMessages.length} / {filteredMessages.length ? currentMessageNumber + 1 : 0}
                         </Text>
-                        <Icon.Button
-                            size={20}
-                            onPress={() => this.prevMessage(currentMessageNumber)}
-                            name="arrow-down"
-                            backgroundColor="d6efef"
-                            color={SOFT_BLUE_COLOR}
-                        />
+                        <Icon.Button size={20} onPress={() => this.prevMessage(currentMessageNumber)} name="arrow-down" backgroundColor="d6efef" color={SOFT_BLUE_COLOR} />
                     </SearchMessagesBar>
                 ) : (
-                    <MessageInput
-                        handleSendMessage={message => this.props.sendMessage(chat.activeChat.chatId, message)}
-                    />
+                    <MessageInput handleSendMessage={message => this.props.sendMessage(chat.activeChat.chatId, message)} />
                 )}
             </ChatView>
         )
